chore: 修改import js示例wasm产物格式
