+++ examples/import-resources/import-js-wasm/index.js
@@ -1,303 +1,5 @@
-import {
-  MyClass,
-  name,
-} from './snippets/import_js-ebef8887557a6fbe/defined-in-js.js';
-
-let wasm;
-
-const cachedTextDecoder =
-  typeof TextDecoder !== 'undefined'
-    ? new TextDecoder('utf-8', { ignoreBOM: true, fatal: true })
-    : {
-        decode: () => {
-          throw Error('TextDecoder not available');
-        },
-      };
-
-if (typeof TextDecoder !== 'undefined') {
-  cachedTextDecoder.decode();
-}
-
-let cachedUint8ArrayMemory0 = null;
-
-function getUint8ArrayMemory0() {
-  if (
-    cachedUint8ArrayMemory0 === null ||
-    cachedUint8ArrayMemory0.byteLength === 0
-  ) {
-    cachedUint8ArrayMemory0 = new Uint8Array(wasm.memory.buffer);
-  }
-  return cachedUint8ArrayMemory0;
-}
-
-function getStringFromWasm0(ptr, len) {
-  ptr = ptr >>> 0;
-  return cachedTextDecoder.decode(
-    getUint8ArrayMemory0().subarray(ptr, ptr + len),
-  );
-}
-
-let WASM_VECTOR_LEN = 0;
-
-const cachedTextEncoder =
-  typeof TextEncoder !== 'undefined'
-    ? new TextEncoder('utf-8')
-    : {
-        encode: () => {
-          throw Error('TextEncoder not available');
-        },
-      };
-
-const encodeString =
-  typeof cachedTextEncoder.encodeInto === 'function'
-    ? function (arg, view) {
-        return cachedTextEncoder.encodeInto(arg, view);
-      }
-    : function (arg, view) {
-        const buf = cachedTextEncoder.encode(arg);
-        view.set(buf);
-        return {
-          read: arg.length,
-          written: buf.length,
-        };
-      };
-
-function passStringToWasm0(arg, malloc, realloc) {
-  if (realloc === undefined) {
-    const buf = cachedTextEncoder.encode(arg);
-    const ptr = malloc(buf.length, 1) >>> 0;
-    getUint8ArrayMemory0()
-      .subarray(ptr, ptr + buf.length)
-      .set(buf);
-    WASM_VECTOR_LEN = buf.length;
-    return ptr;
-  }
-
-  let len = arg.length;
-  let ptr = malloc(len, 1) >>> 0;
-
-  const mem = getUint8ArrayMemory0();
-
-  let offset = 0;
-
-  for (; offset < len; offset++) {
-    const code = arg.charCodeAt(offset);
-    if (code > 0x7f) break;
-    mem[ptr + offset] = code;
-  }
-
-  if (offset !== len) {
-    if (offset !== 0) {
-      arg = arg.slice(offset);
-    }
-    ptr = realloc(ptr, len, (len = offset + arg.length * 3), 1) >>> 0;
-    const view = getUint8ArrayMemory0().subarray(ptr + offset, ptr + len);
-    const ret = encodeString(arg, view);
-
-    offset += ret.written;
-    ptr = realloc(ptr, len, offset, 1) >>> 0;
-  }
-
-  WASM_VECTOR_LEN = offset;
-  return ptr;
-}
-
-let cachedDataViewMemory0 = null;
-
-function getDataViewMemory0() {
-  if (
-    cachedDataViewMemory0 === null ||
-    cachedDataViewMemory0.buffer.detached === true ||
-    (cachedDataViewMemory0.buffer.detached === undefined &&
-      cachedDataViewMemory0.buffer !== wasm.memory.buffer)
-  ) {
-    cachedDataViewMemory0 = new DataView(wasm.memory.buffer);
-  }
-  return cachedDataViewMemory0;
-}
-
-const heap = new Array(128).fill(undefined);
-
-heap.push(undefined, null, true, false);
-
-let heap_next = heap.length;
-
-function addHeapObject(obj) {
-  if (heap_next === heap.length) heap.push(heap.length + 1);
-  const idx = heap_next;
-  heap_next = heap[idx];
-
-  heap[idx] = obj;
-  return idx;
-}
-
-function getObject(idx) {
-  return heap[idx];
-}
-
-function dropObject(idx) {
-  if (idx < 132) return;
-  heap[idx] = heap_next;
-  heap_next = idx;
-}
-
-function takeObject(idx) {
-  const ret = getObject(idx);
-  dropObject(idx);
-  return ret;
-}
-
-export function run() {
-  wasm.run();
-}
-
-async function __wbg_load(module, imports) {
-  if (typeof Response === 'function' && module instanceof Response) {
-    if (typeof WebAssembly.instantiateStreaming === 'function') {
-      try {
-        return await WebAssembly.instantiateStreaming(module, imports);
-      } catch (e) {
-        if (module.headers.get('Content-Type') != 'application/wasm') {
-          console.warn(
-            '`WebAssembly.instantiateStreaming` failed because your server does not serve Wasm with `application/wasm` MIME type. Falling back to `WebAssembly.instantiate` which is slower. Original error:\n',
-            e,
-          );
-        } else {
-          throw e;
-        }
-      }
-    }
-
-    const bytes = await module.arrayBuffer();
-    return await WebAssembly.instantiate(bytes, imports);
-  } else {
-    const instance = await WebAssembly.instantiate(module, imports);
-
-    if (instance instanceof WebAssembly.Instance) {
-      return { instance, module };
-    } else {
-      return instance;
-    }
-  }
-}
-
-function __wbg_get_imports() {
-  const imports = {};
-  imports.wbg = {};
-  imports.wbg.__wbg_log_b2db2c5816f1d07e = function (arg0, arg1) {
-    console.log(getStringFromWasm0(arg0, arg1));
-  };
-  imports.wbg.__wbg_name_fca994005bad39be = function (arg0) {
-    const ret = name();
-    const ptr1 = passStringToWasm0(
-      ret,
-      wasm.__wbindgen_malloc,
-      wasm.__wbindgen_realloc,
-    );
-    const len1 = WASM_VECTOR_LEN;
-    getDataViewMemory0().setInt32(arg0 + 4 * 1, len1, true);
-    getDataViewMemory0().setInt32(arg0 + 4 * 0, ptr1, true);
-  };
-  imports.wbg.__wbg_new_74968321906dc678 = function () {
-    const ret = new MyClass();
-    return addHeapObject(ret);
-  };
-  imports.wbg.__wbg_number_6ea25e36c02a28e3 = function (arg0) {
-    const ret = getObject(arg0).number;
-    return ret;
-  };
-  imports.wbg.__wbg_render_40d583ad633779da = function (arg0, arg1) {
-    const ret = getObject(arg1).render();
-    const ptr1 = passStringToWasm0(
-      ret,
-      wasm.__wbindgen_malloc,
-      wasm.__wbindgen_realloc,
-    );
-    const len1 = WASM_VECTOR_LEN;
-    getDataViewMemory0().setInt32(arg0 + 4 * 1, len1, true);
-    getDataViewMemory0().setInt32(arg0 + 4 * 0, ptr1, true);
-  };
-  imports.wbg.__wbg_setnumber_a86efb097c245512 = function (arg0, arg1) {
-    const ret = (getObject(arg0).number = arg1 >>> 0);
-    return addHeapObject(ret);
-  };
-  imports.wbg.__wbindgen_object_drop_ref = function (arg0) {
-    takeObject(arg0);
-  };
-
-  return imports;
-}
-
-function __wbg_init_memory(imports, memory) {}
-
-function __wbg_finalize_init(instance, module) {
-  wasm = instance.exports;
-  __wbg_init.__wbindgen_wasm_module = module;
-  cachedDataViewMemory0 = null;
-  cachedUint8ArrayMemory0 = null;
-
-  wasm.__wbindgen_start();
-  return wasm;
-}
-
-function initSync(module) {
-  if (wasm !== undefined) return wasm;
-
-  if (typeof module !== 'undefined') {
-    if (Object.getPrototypeOf(module) === Object.prototype) {
-      ({ module } = module);
-    } else {
-      console.warn(
-        'using deprecated parameters for `initSync()`; pass a single object instead',
-      );
-    }
-  }
-
-  const imports = __wbg_get_imports();
-
-  __wbg_init_memory(imports);
-
-  if (!(module instanceof WebAssembly.Module)) {
-    module = new WebAssembly.Module(module);
-  }
-
-  const instance = new WebAssembly.Instance(module, imports);
-
-  return __wbg_finalize_init(instance, module);
-}
-
-async function __wbg_init(module_or_path) {
-  if (wasm !== undefined) return wasm;
-
-  if (typeof module_or_path !== 'undefined') {
-    if (Object.getPrototypeOf(module_or_path) === Object.prototype) {
-      ({ module_or_path } = module_or_path);
-    } else {
-      console.warn(
-        'using deprecated parameters for the initialization function; pass a single object instead',
-      );
-    }
-  }
-
-  if (typeof module_or_path === 'undefined') {
-    module_or_path = new URL('index_bg.wasm', import.meta.url);
-  }
-  const imports = __wbg_get_imports();
-
-  if (
-    typeof module_or_path === 'string' ||
-    (typeof Request === 'function' && module_or_path instanceof Request) ||
-    (typeof URL === 'function' && module_or_path instanceof URL)
-  ) {
-    module_or_path = fetch(module_or_path);
-  }
-
-  __wbg_init_memory(imports);
-
-  const { instance, module } = await __wbg_load(await module_or_path, imports);
-
-  return __wbg_finalize_init(instance, module);
-}
-
-export { initSync };
-export default __wbg_init;
+import * as wasm from './index_bg.wasm';
+export * from './index_bg.js';
+import { __wbg_set_wasm } from './index_bg.js';
+__wbg_set_wasm(wasm);
+wasm.__wbindgen_start();
